Add files via upload
In this example, MyView struct is the main view of the application, it contains a List that displays 
all the ingredients, 
an @ObservedObject property for the Model struct, and an @EnvironmentObject for UserSettings struct. 
Also, the MyView has 
a NavigationView to navigate between views, and a sheet that will be used to add new ingredients and 
it is triggered by a button that appears in the
navigation bar. The `Model` struct is an observable object that contains an array of `Ingredient` structs, 
it uses the `@Published` property wrapper 
to allow the changes in the ingredients array to be observed by the view. Also, the model struct has 
functions to add and delete ingredients, both of these 
functions are using the `context` which is obtained from `@Environment(\.managedObjectContext)` 
to interact with CoreData, and it is saving the changes 
using `saveData()` function.

The `Ingredient` struct conforms to the `NSManagedObject` and `Identifiable` protocols, it is 
used to represent the ingredients stored in CoreData, 
it has properties for id and name.

The `UserSettings` struct is also an observable object that has two `@Published` properties, one of 
they are used to store the new ingredient name, 
and the other one is used to control the visibility of the add ingredient sheet.

`MyViewModifier` is a struct that conforms to the `ViewModifier` protocol, it has a function 
`body` that takes a `Content` view as a parameter and 
returns a modified version of it. In this example, it is modifying the font and the color of 
any view that is using it.

Finally, there's an extension of `View` which add a dynamic function to all views named `myViewModifier()` 
this function is simply returning 
the view after it passed through the `MyViewModifier` struct.

Note that this is a complex example and that the code snippet is incomplete and it may require 
some modifications or additional code to run.

```extension View {
    func myViewModifier() -> some View {
        self.modifier(MyViewModifier())
    }
}```
In this example, I created an extension of View that adds a dynamic function to all views named 
myViewModifier(), this function is simply returning 
the view after it passed through the MyViewModifier struct.
It allows you to add custom behavior or style to any view without modifying the view struct itself, 
making our code more reusable and maintainable.

This example demonstrates how you can use a variety of concepts in SwiftUI and Core Data 
to create a simple app that allows you to manage
 a list of ingredients, for a better understanding of the example, We should take a look at the 
 official documentation for each concept that is used in 
 this example.